refactor(app): extract headless-query-list from headless.ts
